fix(elo): keep fractional wins in animation frames
Elo ties credit 0.5 wins; int() truncated those to 0 in prepare_animation_data.
Serialize wins as float so half-wins from ties survive into the D3 frames.

+++ chapter6/elo-leaderboard/animation.py
@@ -47,7 +47,8 @@ def prepare_animation_data(history_df: pd.DataFrame, top_n: int = 15) -> dict:
                 'name': row.model,
                 'rating': float(row.rating),
                 'matches': int(row.matches),
-                'wins': int(row.wins)
+                # Elo ties credit 0.5 wins; int() would truncate those to 0.
+                'wins': float(row.wins)
             })
         
         frames.append(frame)
